# docs/swagger.yaml
basePath: /api
definitions:
...
  models.AppError:
    properties:
      codigo:
        type: integer
      mensagemErro:
        type: string
    type: object
...
  models.CatalogBatchResult:
    properties:
      itens:
        items:
          $ref: '#/definitions/models.CatalogWriteResult'
        type: array
      quantidade_processada:
        type: integer
    type: object
  models.CatalogWriteResult:
    properties:
      codigo_local:
        type: string
      id:
        type: string
    type: object
...
  models.Page:
    properties:
      items: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  models.ProductBarcodeValue:
    properties:
      ativo:
        type: boolean
      fracao:
        type: number
      principal:
        type: boolean
      valor:
        maxLength: 32
        type: string
    required:
    - ativo
    - principal
    - valor
    type: object
  models.ProductBatchDTO:
    properties:
      itens:
        items:
          $ref: '#/definitions/models.ProductSnapshot'
        maxItems: 1000
        minItems: 1
        type: array
    required:
    - itens
    type: object
  models.ProductCatalogItem:
    properties:
      precos:
        items:
          $ref: '#/definitions/models.ProductPriceSnapshot'
        type: array
      precos_autoritativos:
        type: boolean
      produto:
        $ref: '#/definitions/models.ProductSnapshot'
    type: object
  models.ProductCatalogPage:
    properties:
      itens:
        items:
          $ref: '#/definitions/models.ProductCatalogItem'
        type: array
      limite:
        type: integer
      pagina:
        type: integer
      tem_mais:
        type: boolean
      total:
        type: integer
    type: object
  models.ProductCategoryValue:
    properties:
      codigo_local:
        type: string
      nome:
        maxLength: 100
        type: string
    required:
    - nome
    type: object
  models.ProductPriceBatchDTO:
    properties:
      itens:
        items:
          $ref: '#/definitions/models.ProductPriceSnapshot'
        maxItems: 1000
        minItems: 1
        type: array
    required:
    - itens
    type: object
  models.ProductPriceBatchResult:
    properties:
      itens:
        items:
          $ref: '#/definitions/models.ProductPriceSnapshot'
        type: array
      quantidade_processada:
        type: integer
    type: object
  models.ProductPriceSnapshot:
    properties:
      ativo:
        type: boolean
      codigo_local:
        maxLength: 100
        type: string
      quantidade:
        type: number
      tipo:
        enum:
        - COST
        - SALE
        - WHOLESALE
        type: string
      valido_desde:
        type: string
      valor:
        minimum: 0
        type: number
    required:
    - ativo
    - codigo_local
    - quantidade
    - tipo
    - valor
    type: object
  models.ProductPublicationBatchDTO:
    properties:
      itens:
        items:
          $ref: '#/definitions/models.ProductPublicationDTO'
        maxItems: 500
        minItems: 1
        type: array
    required:
    - itens
    type: object
  models.ProductPublicationBatchResult:
    properties:
      itens:
        items:
          $ref: '#/definitions/models.ProductPublicationResult'
        type: array
      quantidade_processada:
        type: integer
    type: object
  models.ProductPublicationDTO:
    properties:
      codigo_local_candidato:
        description: |-
          RequestNewLocalCode is used when a receiving store found that the
          canonical product code is occupied by another local product. The server
          moves this product mapping atomically to LocalCodeCandidate before it
          fans the publication out, instead of rejecting the whole synchronization.
        type: string
      evento:
        enum:
        - CREATED
        - UPDATED
        - DISABLED
        type: string
      id:
        type: string
      ids_lojas_destino:
        items:
          type: string
        type: array
      precos:
        items:
          $ref: '#/definitions/models.ProductPriceSnapshot'
        type: array
      precos_autoritativos:
        description: |-
          PricesAuthoritative tells the receiver that an empty prices array is
          intentional and must clear the source store's current price tiers.
          Without this bit, an omitted price snapshot is indistinguishable from a
          product publication that simply did not carry pricing.
        type: boolean
      produto:
        $ref: '#/definitions/models.ProductSnapshot'
      solicitar_novo_codigo_local:
        type: boolean
    required:
    - evento
    - produto
    type: object
  models.ProductPublicationResult:
    properties:
      codigo_local:
        type: string
      destinos_enfileirados:
        type: integer
      evento:
        type: string
      hash_payload:
        type: string
      id:
        type: string
      id_produto:
        type: string
      versao:
        type: integer
    type: object
  models.ProductSnapshot:
    properties:
      ativo:
        type: boolean
      cest:
        type: string
      codigo_local:
        maxLength: 100
        type: string
      codigos_barras:
        items:
          $ref: '#/definitions/models.ProductBarcodeValue'
        type: array
      descricao:
        type: string
      grupo:
        $ref: '#/definitions/models.ProductCategoryValue'
      id_produto:
        type: string
      marca:
        type: string
      ncm:
        type: string
      nome:
        maxLength: 255
        type: string
      sku:
        type: string
      subgrupo:
        $ref: '#/definitions/models.ProductCategoryValue'
      unidade:
        type: string
      usa_serial:
        type: boolean
    required:
    - ativo
    - codigo_local
    - nome
    - usa_serial
    type: object
  models.Store:
    properties:
      ativo:
        type: boolean
      cnpj:
        type: string
      codigo:
        type: string
      createdAt:
        type: string
      groupId:
        type: string
      id:
        type: string
      idEmpresa:
        type: string
      nome:
        type: string
      tipo:
        type: string
      updatedAt:
        type: string
    type: object
  models.StoreDTO:
    properties:
      ativo:
        type: boolean
      cnpj:
        maxLength: 20
        type: string
      codigo:
        maxLength: 50
        type: string
      groupId:
        type: string
      nome:
        maxLength: 255
        type: string
      tipo:
        enum:
        - MATRIZ
        - FILIAL
        type: string
    required:
    - codigo
    - groupId
    - nome
    - tipo
    type: object
  models.StoreView:
    properties:
      agentVersion:
        type: string
      ativo:
        type: boolean
      cnpj:
        type: string
      codigo:
        type: string
      createdAt:
        type: string
      customers:
        type: integer
      employees:
        type: integer
      errors:
        type: integer
      groupId:
        type: string
      id:
        type: string
      idEmpresa:
        type: string
      lastSeen:
        type: string
      nome:
        type: string
      pending:
        type: integer
      products:
        type: integer
      status:
        type: string
      suppliers:
        type: integer
      tipo:
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  description: StoreLink integrates stores. Administrators use JWT Bearer; each FNTS_Server
    installation uses its own fixed X-API-Key and database-validated X-Agent-ID.
  title: StoreLink API
  version: "1.0"
paths:
...
  /integracao/v1/precos-produtos/lotes:
    post:
      consumes:
      - application/json
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Product price batch
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ProductPriceBatchDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductPriceBatchResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Create or update up to 1000 store-specific price tiers
      tags:
      - Direct integration
  /integracao/v1/produtos/{codigo_local}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Product code in this store
        in: path
        name: localCode
        required: true
        type: string
      - description: Product snapshot
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ProductSnapshot'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductSnapshot'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Create or update one canonical product
      tags:
      - Direct integration
  /integracao/v1/produtos/{codigo_local}/precos/{tipo}/{quantidade}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Product code in this store
        in: path
        name: localCode
        required: true
        type: string
      - description: 'Price kind: COST, SALE, or WHOLESALE'
        in: path
        name: kind
        required: true
        type: string
      - description: Minimum quantity for this tier
        in: path
        name: quantity
        required: true
        type: number
      - description: Product price tier
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ProductPriceSnapshot'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductPriceSnapshot'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Create or update one store-specific price tier
      tags:
      - Direct integration
  /integracao/v1/produtos/catalogo:
    get:
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Rows per page, up to 500
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductCatalogPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Lê a visão do catálogo da loja autenticada para carga inicial
      tags:
      - Domain integration
  /integracao/v1/produtos/lotes:
    post:
      consumes:
      - application/json
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Product batch
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ProductBatchDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CatalogBatchResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Create or update products in an atomic batch of at most 1000
      tags:
      - Direct integration
  /integracao/v1/produtos/publicacoes:
    post:
      consumes:
      - application/json
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Normalized product publication
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ProductPublicationDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductPublicationResult'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductPublicationResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Publica um produto normalizado e cria mensagens para as lojas do grupo
      tags:
      - Domain integration
  /integracao/v1/produtos/publicacoes/lotes:
    post:
      consumes:
      - application/json
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Normalized product publications
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ProductPublicationBatchDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductPublicationBatchResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Publica até 500 produtos normalizados em uma chamada
      tags:
      - Domain integration
...
  /stores:
    get:
      parameters:
      - description: Número da página
        in: query
        name: page
        type: integer
      - description: Itens por página
        in: query
        name: pageSize
        type: integer
      - description: Busca por nome ou código
        in: query
        name: q
        type: string
      - description: Estado calculado
        in: query
        name: status
        type: string
      - description: UUID do grupo
        in: query
        name: groupId
        type: string
      - description: UUID da loja
        in: query
        name: storeId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Lista lojas com estado recente do agente
      tags:
      - Lojas
    post:
      consumes:
      - application/json
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: Dados da operação
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Store'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Cria loja
      tags:
      - Lojas
  /stores/{id}:
    get:
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: UUID do recurso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StoreView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Consulta detalhes e métricas da loja
      tags:
      - Lojas
...
securityDefinitions:
  AgentAPIKey:
    description: Per-agent fixed key shown once at provisioning or rotation.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Bearer token obtained from /auth/login.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"